fix: move storageservice to own catalog
