--- source/services/index.ts
@@ -1,3 +1,0 @@
-import * as StorageService from "./StorageService";
-
-export { StorageService };
