+++ source/store/AppContext.tsx
@@ -1,8 +1,10 @@
 import React, { useState, useEffect } from "react";
 import env from "react-native-config";
-import StorageService, { APP_ENV_KEY } from "../services/StorageService";
+import StorageService, {
+  APP_ENV_KEY,
+} from "../services/storage/StorageService";
 
-export interface AppProviderState {
+interface AppProviderState {
   mode: string;
   handleSetMode(newMode: string): void;
   isDevMode: boolean;
